This commit was generated by cvs2svn to compensate for changes in r14731, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@14732 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/iecbus.h
@@ -63,7 +63,7 @@ extern BYTE (*iecbus_callback_read)(CLOCK);
 extern void (*iecbus_callback_write)(BYTE, CLOCK);
 
 extern BYTE iecbus_device_read(void);
-extern int  iecbus_device_write(BYTE unit, BYTE data);
+extern int  iecbus_device_write(unsigned int unit, BYTE data);
 extern void (*iecbus_update_ports)(void);
 
 #endif
